Allows hyperlinks in Snapshots: - snapshotEditor now saves raw editorState as JSON rather than HTML - All snapshots related mutation now accept a bodyJson arg and no longer accepts body as an arg
Modifies snapshotBody to show readOnly editor when field bodyJson is
non-null:
- Adds the linkify decorator to the snapshot Body to render links in
the snapshot rather than only in the SnapshotEditor (linkify plugin does
not create a Link Entity that can be saved into the db, so a readOnly
editor is necessary at the moment to render links created by Linkify.

- Checks whether a body field (from old/chrome extension snapshots body)
  or bodyJson exist and renders them accordingly (body field rendered as
  simple HTML via dangeroulsy set HTML prop).
  NOTE: a snapshot should not have both a body and a bodyJson field.
  This has to be kept in mind if editing a snapshot's content is allowed
  in the future. (by default, should both field exist, bodyJson will be
  rendered).

General Note: Body field on snapshots is no longer created in the
mutation, it is only queried. It is also no longer a required field on
the api. This commit will not work without the corresponding commit on
api.